requirements.txt
# Core dependencies
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
httpx==0.23.0
tweepy==4.10.0
pydantic==1.8.2
loguru==0.5.3
redis==4.3.4
pytest==6.2.5
tenacity==8.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==35.0.0

# Monitoring
prometheus-client==0.11.0
